pyproject: remove Python :: 3.13 classifier for now
Mandatory trove-classifiers validation done by setuptools and hatchling
now creates a minefield of FTBFS on distros with old trove-classifiers.

I tried hard to properly solve it, but it's fundamentaly broken upstream
so just drop this for now and do actual development as opposed to tiptoe
around python packaging that's breaking again and again and again for
grotesque reasons.